Adding class to handle image files

+++ include/Sphere.hpp
@@ -10,12 +10,12 @@ public:
 
 	Vector3D center;
 	double radius;
-	int colorR, colorG, colorB;
+	uint8_t colorR, colorG, colorB;
 
 	Sphere();
 	Sphere(Vector3D c);
 	Sphere(Vector3D c, double r);
-	Sphere(Vector3D c, double r, int cR, int cG, int cB);
+	Sphere(Vector3D c, double r, uint8_t cR, uint8_t cG, uint8_t cB);
 };
 
 std::ostream& operator<<(std::ostream& out, const Sphere& s);
